+++ packages/react/src/use-object.ts
@@ -3,7 +3,7 @@ import {
   isDeepEqualData,
   parsePartialJson,
 } from '@ai-sdk/ui-utils';
-import { useId, useState } from 'react';
+import { useCallback, useId, useRef, useState } from 'react';
 import useSWR from 'swr';
 import z from 'zod';
 
@@ -33,9 +33,14 @@ export type Experimental_UseObjectOptions<RESULT> = {
 
 export type Experimental_UseObjectHelpers<RESULT, INPUT> = {
   /**
-   * Calls the API with the provided input as JSON body.
+   * @deprecated Use `submit` instead.
    */
   setInput: (input: INPUT) => void;
+
+  /**
+   * Calls the API with the provided input as JSON body.
+   */
+  submit: (input: INPUT) => void;
 
   /**
    * The current value for the generated object. Updated as the API streams JSON chunks.
@@ -51,6 +56,11 @@ export type Experimental_UseObjectHelpers<RESULT, INPUT> = {
    * Flag that indicates whether an API request is in progress.
    */
   isLoading: boolean;
+
+  /**
+   * Abort the current request immediately, keep the current partial object if any.
+   */
+  stop: () => void;
 };
 
 function useObject<RESULT, INPUT = any>({
@@ -76,58 +86,75 @@ function useObject<RESULT, INPUT = any>({
   const [error, setError] = useState<undefined | unknown>(undefined);
   const [isLoading, setIsLoading] = useState(false);
 
-  return {
-    async setInput(input) {
-      try {
-        setIsLoading(true);
+  // Abort controller to cancel the current API call.
+  const abortControllerRef = useRef<AbortController | null>(null);
 
-        const response = await fetch(api, {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify(input),
-        });
+  const stop = useCallback(() => {
+    abortControllerRef.current?.abort();
+    abortControllerRef.current = null;
+  }, []);
 
-        if (!response.ok) {
-          throw new Error(
-            (await response.text()) ?? 'Failed to fetch the response.',
-          );
-        }
+  const submit = async (input: INPUT) => {
+    try {
+      setIsLoading(true);
 
-        if (response.body == null) {
-          throw new Error('The response body is empty.');
-        }
+      const abortController = new AbortController();
+      abortControllerRef.current = abortController;
 
-        let accumulatedText = '';
-        let latestObject: DeepPartial<RESULT> | undefined = undefined;
+      const response = await fetch(api, {
+        method: 'POST',
+        headers: { 'Content-Type': 'application/json' },
+        signal: abortController.signal,
+        body: JSON.stringify(input),
+      });
 
-        response.body.pipeThrough(new TextDecoderStream()).pipeTo(
-          new WritableStream<string>({
-            write(chunk) {
-              accumulatedText += chunk;
-
-              const currentObject = parsePartialJson(
-                accumulatedText,
-              ) as DeepPartial<RESULT>;
-
-              if (!isDeepEqualData(latestObject, currentObject)) {
-                latestObject = currentObject;
-
-                mutate(currentObject);
-              }
-            },
-          }),
+      if (!response.ok) {
+        throw new Error(
+          (await response.text()) ?? 'Failed to fetch the response.',
         );
-
-        setError(undefined);
-      } catch (error) {
-        setError(error);
-      } finally {
-        setIsLoading(false);
       }
-    },
+
+      if (response.body == null) {
+        throw new Error('The response body is empty.');
+      }
+
+      let accumulatedText = '';
+      let latestObject: DeepPartial<RESULT> | undefined = undefined;
+
+      response.body.pipeThrough(new TextDecoderStream()).pipeTo(
+        new WritableStream<string>({
+          write(chunk) {
+            accumulatedText += chunk;
+
+            const currentObject = parsePartialJson(
+              accumulatedText,
+            ) as DeepPartial<RESULT>;
+
+            if (!isDeepEqualData(latestObject, currentObject)) {
+              latestObject = currentObject;
+
+              mutate(currentObject);
+            }
+          },
+        }),
+      );
+
+      setError(undefined);
+    } catch (error) {
+      setError(error);
+    } finally {
+      setIsLoading(false);
+      abortControllerRef.current = null;
+    }
+  };
+
+  return {
+    setInput: submit, // Deprecated
+    submit,
     object: data,
     error,
     isLoading,
+    stop,
   };
 }