+++ packages/react/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @ai-sdk/react
 
+## 0.0.11
+
+### Patch Changes
+
+- 827ef450: feat (ai/ui): improve error handling in useAssistant
+
+## 0.0.10
+
+### Patch Changes
+
+- 5b04204b: chore (ai/react): rename useChat setInput to submit
+- 8f482903: feat (ai/react): add stop() helper to useObject
+
+## 0.0.9
+
+### Patch Changes
+
+- 321a7d0e: feat (ai/react): add isLoading to useObject
+
 ## 0.0.8
 
 ### Patch Changes